New circular buffer for streaming. Seems to send metrics but no debug shown??
