+++ database/rrd.h
@@ -703,11 +703,15 @@ struct rrdhost {
 
     SIMPLE_PATTERN *rrdpush_send_charts_matching;   // pattern to match the charts to be sent
 
+    // Metric transmission: collector threads asynchronously fill the buffer, sender thread uses it.
+    struct circular_buffer *sender_buffer;
+    BUFFER *sender_build;
+
     // metrics may be collected asynchronously
     // these synchronize all the threads willing the write to our sending buffer
     netdata_mutex_t rrdpush_sender_buffer_mutex;    // exclusive access to rrdpush_sender_buffer
     int rrdpush_sender_pipe[2];                     // collector to sender thread signaling
-    BUFFER *rrdpush_sender_buffer;                  // collector fills it, sender sends it
+    //BUFFER *rrdpush_sender_buffer;                  // collector fills it, sender sends it
 
     uint32_t stream_version;                             //Set the current version of the stream.
